+++ test.js
@@ -126,6 +126,19 @@ function noop() {
             "/*jslint name*/\nlet aa = aa._;",
             "/*jslint single*/\nlet aa = 'aa';",
             "/*jslint this*/\nlet aa = this;",
+            (
+                "/*jslint unordered*/\n"
+                + "function aa({bb, aa}) {\n"
+                + "    switch (aa) {\n"
+                + "    case 1:\n"
+                + "        break;\n"
+                + "    case 0:\n"
+                + "        break;\n"
+                + "    default:\n"
+                + "        return {bb, aa};\n"
+                + "    }\n"
+                + "}\n"
+            ),
             "/*jslint unordered*/\nlet {bb, aa} = 0;",
             "/*jslint white*/\n\t",
             "/*property aa bb*/"
@@ -269,6 +282,8 @@ function noop() {
             + "|stop_at"
             + "|warn"
             + "|warn_at"
+            + "|warn_if_unordered"
+            + "|warn_if_unordered_case_statement"
             + ")"
             + "\\\u0028\\s*?\"?"
             + "(\\S[^\n\"]+)"
@@ -293,6 +308,10 @@ function noop() {
                 break;
             case "semicolon":
                 expectedWarningCode = "expected_a_b";
+                break;
+            case "warn_if_unordered":
+            case "warn_if_unordered_case_statement":
+                expectedWarningCode = "expected_a_b_before_c_d";
                 break;
             }
         }